requirement.txt
Django==1.11.4
gitdb2==2.0.2
GitPython==2.1.5
image==1.5.11
olefile==0.44
Pillow==4.2.1
pytz==2017.2
smmap2==2.0.3
